# Conflicts:
#	packages/database/src/repositories/mod.rs

+++ packages/database/src/repositories/author_repository.rs
@@ -5,7 +5,7 @@ use std::sync::Arc;
 
 #[async_trait]
 pub trait AuthorRepo {
-    async fn get_author(&self, id: i32) -> anyhow::Result<Author>;
+    async fn get_author(&self, id: String) -> anyhow::Result<Author>;
     async fn add_author(&self, author: Author) -> anyhow::Result<()>;
     async fn edit_author(&self, author: Author) -> anyhow::Result<()>;
     async fn delete_author(&self, author: Author) -> anyhow::Result<()>;
@@ -23,12 +23,8 @@ impl AuthorRepository {
 
 #[async_trait]
 impl AuthorRepo for AuthorRepository {
-    async fn get_author(&self, id: i32) -> anyhow::Result<Author> {
-        let author = sqlx::query_as!(
-            Author,
-            "SELECT * FROM author WHERE id = ?",
-            author.id,
-        )
+    async fn get_author(&self, id: String) -> anyhow::Result<Author> {
+        let author = sqlx::query_as!(Author, "SELECT * FROM author as a WHERE a.id = ?", id)
             .fetch_one(&*self.mysql_pool)
             .await?;
 
@@ -37,8 +33,7 @@ impl AuthorRepo for AuthorRepository {
 
     async fn add_author(&self, author: Author) -> anyhow::Result<()> {
         sqlx::query!(
-            "INSERT INTO author (id, name)
-             VALUES (?, ?)",
+            "INSERT INTO author (id, name) VALUES(?, ?)",
             author.id,
             author.name
         )
@@ -50,29 +45,23 @@ impl AuthorRepo for AuthorRepository {
 
     async fn edit_author(&self, author: Author) -> anyhow::Result<()> {
         sqlx::query!(
-            "UPDATE author SET (name)
-             VALUES (?) WHERE id = (?)",
+            "UPDATE author SET name = ?
+            WHERE id = ?",
             author.name,
             author.id
         )
-            .execute(&*self.mysql_pool)
-            .await?;
+        .execute(&*self.mysql_pool)
+        .await?;
 
         Ok(())
     }
 
     async fn delete_author(&self, author: Author) -> anyhow::Result<()> {
-        sqlx::query!(
-            "DELETE FROM author_book WHERE author_id = (?)",
-            author.id,
-        )
+        sqlx::query!("DELETE FROM author_book WHERE author_id = ?", author.id)
             .execute(&*self.mysql_pool)
             .await?;
 
-        sqlx::query!(
-            "DELETE FROM author WHERE id = (?)",
-            author.id,
-        )
+        sqlx::query!("DELETE FROM author WHERE id = (?)", author.id,)
             .execute(&*self.mysql_pool)
             .await?;
